Implement AC.1 shared storage contract
