+++ crates/atm-core/Cargo.toml
@@ -20,6 +20,7 @@ name = "atm_core"
 test-utils = []
 
 [dependencies]
+atm-storage = { path = "../atm-storage", version = "1.2.1" }
 serde.workspace = true
 serde_json.workspace = true
 tracing.workspace = true
