+++ crates/atm-core/src/schema/mod.rs
@@ -6,6 +6,7 @@ pub mod permissions;
 pub mod settings;
 pub mod team_config;
 
-pub use agent_member::{AgentMember, AgentType};
-pub use inbox_message::{AtmMessageId, MessageEnvelope, PendingAck, ThreadMode};
+pub use agent_member::AgentMember;
+pub use atm_storage::contract::AgentType;
+pub use inbox_message::{AlertKind, AtmMessageId, MessageEnvelope, PendingAck, ThreadMode};
 pub use team_config::TeamConfig;
