+++ crates/atm-storage/src/contract.rs
@@ -1,0 +1,313 @@
+use serde::{Deserialize, Serialize};
+use serde_json::{Map, Value};
+use std::fmt;
+use std::str::FromStr;
+
+use crate::error::AtmError;
+use crate::schema::{AtmMessageId, MessageEnvelope};
+use crate::types::{AgentName, IsoTimestamp, ModelName, PaneId, TaskId, TeamName};
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, PartialOrd, Ord, Hash)]
+#[serde(transparent)]
+pub struct MessageKey(String);
+
+impl MessageKey {
+    pub fn new(value: impl Into<String>) -> Result<Self, AtmError> {
+        let value = value.into();
+        if value.trim().is_empty() {
+            return Err(
+                AtmError::validation("message key must not be blank").with_recovery(
+                    "Populate a stable ATM message key before calling the storage contract.",
+                ),
+            );
+        }
+        Ok(Self(value))
+    }
+
+    pub fn into_inner(self) -> String {
+        self.0
+    }
+
+    pub fn as_str(&self) -> &str {
+        &self.0
+    }
+
+    pub fn as_atm_message_id(&self) -> Result<AtmMessageId, AtmError> {
+        let raw = self.as_str().strip_prefix("atm:").unwrap_or(self.as_str());
+        raw.parse::<AtmMessageId>()
+            .map_err(|error| AtmError::validation(format!("message key parse failed: {error}")))
+    }
+}
+
+impl From<AtmMessageId> for MessageKey {
+    fn from(value: AtmMessageId) -> Self {
+        Self(format!("atm:{value}"))
+    }
+}
+
+impl FromStr for MessageKey {
+    type Err = AtmError;
+
+    fn from_str(value: &str) -> Result<Self, Self::Err> {
+        Self::new(value)
+    }
+}
+
+impl fmt::Display for MessageKey {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        f.write_str(self.as_str())
+    }
+}
+
+impl AsRef<str> for MessageKey {
+    fn as_ref(&self) -> &str {
+        self.as_str()
+    }
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, PartialOrd, Ord, Hash)]
+#[serde(transparent)]
+pub struct TaskState(String);
+
+impl TaskState {
+    pub fn new(value: impl Into<String>) -> Result<Self, AtmError> {
+        let value = value.into();
+        if value.trim().is_empty() {
+            return Err(
+                AtmError::validation("task state must not be blank").with_recovery(
+                    "Populate a non-empty task state before calling the storage contract.",
+                ),
+            );
+        }
+        Ok(Self(value))
+    }
+}
+
+impl AsRef<str> for TaskState {
+    fn as_ref(&self) -> &str {
+        &self.0
+    }
+}
+
+impl fmt::Display for TaskState {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        f.write_str(self.as_ref())
+    }
+}
+
+impl FromStr for TaskState {
+    type Err = AtmError;
+
+    fn from_str(value: &str) -> Result<Self, Self::Err> {
+        Self::new(value)
+    }
+}
+
+impl PartialEq<&str> for TaskState {
+    fn eq(&self, other: &&str) -> bool {
+        self.as_ref() == *other
+    }
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, PartialOrd, Ord, Hash)]
+#[serde(transparent)]
+pub struct AckTransition(String);
+
+impl AckTransition {
+    pub fn new(value: impl Into<String>) -> Result<Self, AtmError> {
+        let value = value.into();
+        if value.trim().is_empty() {
+            return Err(
+                AtmError::validation("ack transition must not be blank").with_recovery(
+                    "Populate a non-empty ack transition before calling the storage contract.",
+                ),
+            );
+        }
+        Ok(Self(value))
+    }
+}
+
+impl AsRef<str> for AckTransition {
+    fn as_ref(&self) -> &str {
+        &self.0
+    }
+}
+
+impl fmt::Display for AckTransition {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        f.write_str(self.as_ref())
+    }
+}
+
+impl FromStr for AckTransition {
+    type Err = AtmError;
+
+    fn from_str(value: &str) -> Result<Self, Self::Err> {
+        Self::new(value)
+    }
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
+pub struct Message {
+    pub team: TeamName,
+    pub agent: AgentName,
+    pub message_key: MessageKey,
+    pub envelope: MessageEnvelope,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+pub struct MessageQuery {
+    pub team: TeamName,
+    pub agent: AgentName,
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub sender: Option<AgentName>,
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub task_id: Option<TaskId>,
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub limit: Option<usize>,
+}
+
+#[derive(Debug, Clone, Copy, Serialize, Deserialize, PartialEq, Eq)]
+#[serde(rename_all = "kebab-case")]
+pub enum RosterMemberKind {
+    Permanent,
+    Ephemeral,
+}
+
+#[derive(Debug, Clone, Copy, Serialize, Deserialize, PartialEq, Eq)]
+#[serde(rename_all = "kebab-case")]
+pub enum RosterHarness {
+    ClaudeCode,
+    CodexCli,
+    GeminiCli,
+    Opencode,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub enum AgentType {
+    GeneralPurpose,
+    Plan,
+    Lead,
+    Qa,
+    Worker,
+    Unknown(String),
+}
+
+impl Default for AgentType {
+    fn default() -> Self {
+        Self::Unknown(String::new())
+    }
+}
+
+impl From<String> for AgentType {
+    fn from(value: String) -> Self {
+        match value.as_str() {
+            "general-purpose" => Self::GeneralPurpose,
+            "plan" => Self::Plan,
+            "lead" => Self::Lead,
+            "qa" => Self::Qa,
+            "worker" => Self::Worker,
+            _ => Self::Unknown(value),
+        }
+    }
+}
+
+impl AgentType {
+    pub fn as_str(&self) -> &str {
+        match self {
+            Self::GeneralPurpose => "general-purpose",
+            Self::Plan => "plan",
+            Self::Lead => "lead",
+            Self::Qa => "qa",
+            Self::Worker => "worker",
+            Self::Unknown(value) => value,
+        }
+    }
+}
+
+impl From<AgentType> for String {
+    fn from(value: AgentType) -> Self {
+        value.as_str().to_string()
+    }
+}
+
+impl Serialize for AgentType {
+    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: serde::Serializer,
+    {
+        serializer.serialize_str(self.as_str())
+    }
+}
+
+impl fmt::Display for AgentType {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        f.write_str(self.as_str())
+    }
+}
+
+impl<'de> Deserialize<'de> for AgentType {
+    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+    where
+        D: serde::Deserializer<'de>,
+    {
+        Ok(Self::from(String::deserialize(deserializer)?))
+    }
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+pub struct RosterMember {
+    pub team_name: TeamName,
+    pub agent_name: AgentName,
+    pub member_kind: RosterMemberKind,
+    pub harness: RosterHarness,
+    #[serde(default)]
+    pub agent_type: AgentType,
+    #[serde(default)]
+    pub model: ModelName,
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub recipient_pane_id: Option<PaneId>,
+    #[serde(default)]
+    pub metadata_json: Map<String, Value>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+pub struct RosterSnapshot {
+    pub team_name: TeamName,
+    pub members: Vec<RosterMember>,
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub refreshed_at: Option<IsoTimestamp>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+pub struct MessageReceivedEvent {
+    pub team: TeamName,
+    pub agent: AgentName,
+    pub message_key: MessageKey,
+    pub timestamp: IsoTimestamp,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+pub struct RosterChangedEvent {
+    pub team: TeamName,
+    pub member_count: usize,
+    pub timestamp: IsoTimestamp,
+}
+
+pub trait MessageStore {
+    fn save_message(&self, message: &Message) -> Result<(), AtmError>;
+    fn load_message(&self, key: &MessageKey) -> Result<Option<Message>, AtmError>;
+    fn list_messages(&self, query: &MessageQuery) -> Result<Vec<Message>, AtmError>;
+    fn delete_message(&self, key: &MessageKey) -> Result<(), AtmError>;
+}
+
+pub trait RosterStore {
+    fn load_roster(&self, team: &TeamName) -> Result<RosterSnapshot, AtmError>;
+    fn save_roster(&self, roster: &RosterSnapshot) -> Result<(), AtmError>;
+    fn list_teams(&self) -> Result<Vec<TeamName>, AtmError>;
+}
+
+pub trait StorageNotifier {
+    fn message_received(&self, event: &MessageReceivedEvent) -> Result<(), AtmError>;
+    fn roster_changed(&self, event: &RosterChangedEvent) -> Result<(), AtmError>;
+}
