+++ crates/atm-storage/src/schema/mod.rs
@@ -1,0 +1,4 @@
+mod inbox_message;
+
+pub use inbox_message::{AlertKind, AtmMessageId, MessageEnvelope, PendingAck, ThreadMode};
+
