+++ crates/atm-storage/src/validation.rs
@@ -1,0 +1,39 @@
+use crate::error::AtmError;
+
+pub(crate) fn validate_path_segment(value: &str, kind: &str) -> Result<(), AtmError> {
+    if value.is_empty() {
+        return Err(AtmError::address_parse(format!(
+            "{kind} name must not be empty"
+        )));
+    }
+
+    if value.starts_with('.') {
+        return Err(AtmError::address_parse(format!(
+            "{kind} name must not start with '.'"
+        )));
+    }
+
+    if value.contains("..") {
+        return Err(AtmError::address_parse(format!(
+            "{kind} name must not contain '..'"
+        )));
+    }
+
+    if value.contains(['/', '\\']) {
+        return Err(AtmError::address_parse(format!(
+            "{kind} name must not contain path separators"
+        )));
+    }
+
+    if !value
+        .chars()
+        .all(|ch| ch.is_ascii_alphanumeric() || matches!(ch, '-' | '_' | '.'))
+    {
+        return Err(AtmError::address_parse(format!(
+            "{kind} name contains invalid characters"
+        )));
+    }
+
+    Ok(())
+}
+
