+++ docs/plans/phase-AC/readiness.md
@@ -17,7 +17,7 @@ Authoritative supporting inventory:
 | Sprint | Status | Branch | Worktree | Closure Gate |
 | --- | --- | --- | --- | --- |
 | `AC.0` | `complete` | `plan/phase-AC` | `../atm-core-worktrees/plan/phase-AC` | ADR-018 and the Phase AC issue inventory freeze the original design reset and the target crate graph before execution sprint planning continues |
-| `AC.1` | `planned` | `feature/pAC-s1-atm-storage-contract-and-canonical-types` | `../atm-core-worktrees/feature/pAC-s1-atm-storage-contract-and-canonical-types` | `crates/atm-storage` lands with a small audited message/roster contract, canonical domain structs, and separate notification traits |
+| `AC.1` | `complete` | `feature/pAC-s1-atm-storage-contract-and-canonical-types` | `../atm-core-worktrees/feature/pAC-s1-atm-storage-contract-and-canonical-types` | `crates/atm-storage` landed with a small audited message/roster contract, canonical shared message/roster structs, separate notification traits, and an `atm-core` compile bridge |
 | `AC.2` | `planned` | `feature/pAC-s2-atm-storage-claude-extraction` | `../atm-core-worktrees/feature/pAC-s2-atm-storage-claude-extraction` | Claude inbox storage is extracted behind the shared traits without widening the contract to file-format specifics |
 | `AC.3` | `planned` | `feature/pAC-s3-sqlite-backend-convergence` | `../atm-core-worktrees/feature/pAC-s3-sqlite-backend-convergence` | the SQLite backend implements the same approved mail/roster contract, emits notifications only after durable write success, and no longer depends on `atm-core` |
 | `AC.4` | `planned` | `feature/pAC-s4-atm-core-storage-boundary-adoption` | `../atm-core-worktrees/feature/pAC-s4-atm-core-storage-boundary-adoption` | `atm-core`, runtime, and daemon paths depend on storage traits rather than concrete backend seams |
