+++ docs/plans/phase-AC/sprint-AC1.md
@@ -6,7 +6,7 @@ phase: AC
 sprint: AC.1
 worktree: ../atm-core-worktrees/feature/pAC-s1-atm-storage-contract-and-canonical-types
 branch: feature/pAC-s1-atm-storage-contract-and-canonical-types
-status: planned
+status: complete
 estimated_scope: large
 ```
 
@@ -246,3 +246,10 @@ Proof this sprint must leave behind:
 - if the contract copies the current boundary request/response volume, the phase has failed
 - if the shared structs are still transport-shaped rather than semantic, `AC.5` will stall
 - if this sprint leaves both old and new shared contracts standing indefinitely, `AC.4` and `AC.6` will false-close
+
+## Closure Note
+
+- `AC.1` landed `crates/atm-storage` with canonical message/roster contract
+  traits, canonical shared message/roster structs, notification events, and an
+  `atm-core` compile bridge that keeps the workspace buildable while later
+  consumer-migration sprints close the legacy surfaces
